requirements.txt
Flask
openai
azure-identity
gunicorn